Picks Random Song
Script picks random song from an array and puts it into the audio source.

Then the audio source takes the song and grabs the audio spectrum data.

The audioVisualize script takes the spectrum data and instiniates cubes in a 360 circle. then each frame the script retakes the spectrum data and updates the cubes transform.

When the song is over the audio script picks another random song from the list